+++ tests/test_utils.py
@@ -59,14 +59,14 @@ class TestUtils(unittest.TestCase):
         qs_beyond_val = 400
         qs_beyoud_arr = [-50, 200, 400]
         zq = pd.DataFrame({
-            'name' : [name] * 3,
-            'Q' : [0, 200, 300],
-            'Z' : [0, 100, 200]
+            'station_id' : [name] * 3,
+            'discharge' : [0, 200, 300],
+            'tailrace_level' : [0, 100, 200]
         })
         zs = pd.DataFrame({
-            'name' : [name] * 3,
-            'V' : [0, 200, 300],
-            'Z' : [0, 100, 200]
+            'station_id' : [name] * 3,
+            'volume' : [0, 200, 300],
+            'forebay_level' : [0, 100, 200]
         })
         self.assertAlmostEqual(
             interpolate_z_by_q_or_s(name, qs_val, zq),